project/python-service/requirements.txt
# requirements-tft.txt   ──≈ 350 MB
torch==2.1.2+cpu          --extra-index-url https://download.pytorch.org/whl/cpu
pytorch-lightning<2.2
pytorch-forecasting
scipy
fastapi
uvicorn[standard]
boto3
numpy==1.26.*
pandas
dtaidistance
